Refactors microSWIFT.py and identifies code to pull out
