+++ microSWIFT/utils/utils.py
@@ -1,7 +1,12 @@
-def fillBadValues(badVal=999,spectralLen=42):
+""" 
+Catch-all for general utilities.
+"""
+
+import numpy as np
+
+
+def fill_bad_values(badVal=999, spectralLen=42):
     """
-    Author: @jacobrdavis
-    
     Utility to fill wave parameters with bad values.
 
     Input:
@@ -28,8 +33,6 @@ def fillBadValues(badVal=999,spectralLen=42):
     Example:
         u,v,z,lat,lon,Hs,Tp,Dp,E,f,a1,b1,a2,b2,check = fillBadValues(badVal=999,spectralLen=42)
     """
-    #--Import Statements
-    import numpy as np
 
     u     = badVal
     v     = badVal
@@ -49,4 +52,21 @@ def fillBadValues(badVal=999,spectralLen=42):
 
     return u,v,z,lat,lon,Hs,Tp,Dp,E,f,a1,b1,a2,b2,check
 
+def get_uvzmean(badValue, pts):
+    """TODO:"""
+    mean = badValue #set values to 999 initially and fill if valid value
+    index = np.where(pts != badValue)[0] #get index of non bad values
+    pts=pts[index] #take subset of data without bad values in it
 
+    if(len(index) > 0):
+        mean = np.mean(pts)
+
+    return mean
+
+def get_last(badValue, pts):
+    """TODO:"""
+    for i in range(1, len(pts)): #loop over entire lat/lon array
+        if pts[-i] != badValue: #count back from last point looking for a real position
+            return pts[-i]
+        
+    return badValue #returns badValue if no real position exists
